requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.6
django-filter==22.1
django-rest-knox==1.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dynaconf==3.1.9
gunicorn==20.1.0
hvac==0.11.2
idna==3.3
knox==0.1.14
loguru==0.6.0
Markdown==3.4.1
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pytz==2022.1
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
validators==0.20.0
whitenoise==6.2.0
